requirements/base.txt
Django==5.2.4
python-environ==0.4.54
psycopg2-binary==2.9.10
django-ckeditor-5==0.2.18
django-mptt==0.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0